Allow custom non-damage weaknesses, fix Blessed Counterstrike weakness (#19183)
This allows creating custom non-damage weaknesses.  E.g., a definition of
"origin:signature:{item|origin.signature}" would create a weakness to strikes
from a specific player who is the origin of the effect with the weakness.  By
setting applyOnce to true, the weakness will only trigger once for a damage
roll, rather than triggering on every damage instance in the roll.